guard library tab for non-nostr users
