+++ components/index.ts
@@ -71,3 +71,4 @@ export * from "./TopUpSection";
 export * from "./ForYouSection";
 export * from "./FeaturedSection";
 export * from "./ProfilePicturePicker";
+export * from "./AddNostr";
